Fix config saving:
- Use plugin for data structure
- Remove group plugin
- Clean up DI
