+++ etc/adminhtml/di.xml
@@ -19,7 +19,7 @@
  -->
 
 <config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
-    <type name="\Magento\Config\Model\Config\Structure\Element\Group">
-        <plugin name="dynamic_config_fields" type="Magenerds\CountryPopUp\Plugin\Config\Group"/>
+    <type name="Magento\Config\Model\Config\Structure\Data">
+        <plugin name="config_structure_data_plugin" type="Magenerds\CountryPopUp\Plugin\Config\Structure\DataPlugin"/>
     </type>
 </config>
